Fixed minor bugs
