+++ backend/app/database/base.py
@@ -2,4 +2,4 @@ from app.database.base_class import Base
 from app.models.admin import Admin
 from app.models.currency import Currency
 from app.models.rate import Rate
-from app.models.otp import Otp, Otp_Blocks
+
