+++ backend/app/models/admin.py
@@ -1,5 +1,4 @@
 from sqlalchemy import Column, Integer, String, Boolean
-
 from app.database.base_class import Base
 
 
@@ -10,3 +9,5 @@ class Admin(Base):
     email = Column(String, unique=True, index=True)
     password = Column(String)
     is_active = Column(Boolean, default=True)
+    token = Column(String, default="")
+
