+++ backend/app/schemas/admin.py
@@ -14,6 +14,7 @@ class AdminCreate(AdminBase):
 class Admin(AdminBase):
     id: int
     is_active: bool
+    token: str
 
     class Config:
         orm_mode = True
